feat(experimental): mohist remapper testing successfully
